requirements.txt
Django==6.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
matplotlib==3.8.2
plotly==5.18.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
